adapt tool with problem's of skygserv

+++ replace_text_main.py
@@ -26,6 +26,7 @@ logging.basicConfig(level=logging.DEBUG,
                     format='%(asctime)s:%(levelname)s:%(message)s')
 
 csv_file_path = '/home/nhulq/Desktop/shutto/tool-replace-text/input_folder/input.csv'
+csv_file_path_test = '/home/nhulq/Desktop/shutto/tool-replace-text/input_folder/input_test.csv'
 csv_file_output_path = '/home/nhulq/Desktop/shutto/tool-replace-text/output_folder/output.csv'
 folder_need_to_change_path = '/home/nhulq/Desktop/airtrip/skygserv/web'
 
@@ -87,54 +88,66 @@ def replace_text_in_files_of_folder(folder_target_path=None, old_text=None, new_
     _replace_reg = re.compile(old_text)
     for dir_path, dir_names, file_names in os.walk(folder_target_path):
 
+
         # loop to all files in folder
         for file in file_names:
+            is_found = False
+            if file.endswith(".jsp") or file.endswith(".vm"):
+                file = os.path.join(dir_path, file)  # concat filename to path (/home/nhulq/index.html)
+                target_txt_file = file + ".txt"
+                try:
+                    with open(target_txt_file, "w", encoding="shift-jis") as target_file:  # open new file to write
+                        with open(file, errors="ignore", encoding="shift-jis") as source_file:  # open file in OS to find old_text!
+                            # loop each line in current file
+                            for line_text in source_file:
+                                find_result = line_text.find(old_text)
 
-            file = os.path.join(dir_path, file)  # concat filename to path (/home/nhulq/index.html)
-            target_txt_file = file + ".txt"
-            try:
-                with open(target_txt_file, "w", encoding="shift-jis") as target_file:  # open new file to write
-                    with open(file, errors="ignore", encoding="shift-jis") as source_file:  # open file in OS to find old_text!
-                        # loop each line in current file
-                        for line_text in source_file:
-                            line_is_changed = _replace_reg.sub(new_text,
-                                                               line_text)  # return text (find and replace old_text become new_text in line)
-                            if line_text != line_is_changed:  # found at current line!
-                                logging.debug('Đã tìm thấy: \'' + old_text + '\' tại file: ' + file)
-                                log_count_number_found_old_text += 1
-                                log_found_files_list.append(file)
+                                # line_is_changed = _replace_reg.sub(new_text,
+                                #                                    line_text)  # return text (find and replace old_text become new_text in line)
+                                # if line_text != line_is_changed:  # found at current line!
+                                if find_result != -1: # rewrite file
+                                    is_found = True
+                                    logging.debug('Đã tìm thấy: \'' + old_text + '\' tại file: ' + file)
+                                    log_count_number_found_old_text += 1
+                                    log_found_files_list.append(file)
 
-                                if new_text != 'delete':
-                                    target_file.write(line_is_changed)
+                                    if new_text != 'delete':
+                                        result_replace = line_text.replace(old_text, new_text)
+                                        target_file.write(result_replace)
 
-                                    logging.debug('Change thành công')
-                                    log_count_number_changed += 1
-                                else:
-                                    if html_tag_type == 'img' or html_tag_type == 'link':  # action to delete
-                                        target_file.write(remove_html_tags(line_text, html_tag_type))
-
-                                        logging.debug('Xóa thành công thẻ: ' + html_tag_type)
+                                        logging.debug('Change thành công')
                                         log_count_number_changed += 1
-
-                                    elif html_tag_type == 'script' or html_tag_type == 'SCRIPT':
-                                        result = remove_html_tags(line_text, html_tag_type)
-                                        if result != False:
-                                            target_file.write(result)
+                                    else:
+                                        if html_tag_type == 'img' or html_tag_type == 'link':  # action to delete
+                                            target_file.write(remove_html_tags(line_text, html_tag_type))
 
                                             logging.debug('Xóa thành công thẻ: ' + html_tag_type)
                                             log_count_number_changed += 1
-                                        else:  # can't delete whole script tag, because not found </script> tag
-                                            logging.debug('Script tag failed: ' + line_text)
-                                            logging.debug('Xóa thất bại thẻ ' + html_tag_type + ' tại file: ' + file)
-                                            log_number_failed += 1
 
-                                            target_file.write(line_text)
+                                        elif html_tag_type == 'script' or html_tag_type == 'SCRIPT':
+                                            result = remove_html_tags(line_text, html_tag_type)
+                                            if result != False:
+                                                target_file.write(result)
 
-                            else:  # not found old_text
-                                target_file.write(line_text)
-            finally:
-                os.remove(file)  # apply for Windows OS, let's comment this in linux OS
-                os.rename(target_txt_file, file)
+                                                logging.debug('Xóa thành công thẻ: ' + html_tag_type)
+                                                log_count_number_changed += 1
+                                            else:  # can't delete whole script tag, because not found </script> tag
+                                                logging.debug('Script tag failed: ' + line_text)
+                                                logging.debug('Xóa thất bại thẻ ' + html_tag_type + ' tại file: ' + file)
+                                                log_number_failed += 1
+
+                                                target_file.write(line_text)
+
+                                else:  # not found old_text
+                                    target_file.write(line_text)
+
+                finally:
+                    if is_found == False:
+                        os.remove(target_txt_file)# remove file not contain old_text
+                    else:
+                        os.remove(file)  # apply for Windows OS, let's comment this in linux OS
+                        os.rename(target_txt_file, file)
+
 
     if log_count_number_found_old_text == 0:
         logging.error('Không tìm thấy bất kỳ file nào. xin check lại!')
@@ -198,7 +211,9 @@ def replace_text_of_folder(csv_file_path, folder_path, html_tag_type):
 
 
 if __name__ == '__main__':
-    # p = remove_html_tags('<script src="../jquery/jquery-1.6.1.min.js"><link src="sdas">', 'script')
+    # p = remove_html_tags('<link rel="stylesheet" href="/css/common.css" media="screen"> '
+    #                      '<link rel="stylesheet" href="/css/cc.css" media="screen">'
+    #                      '<link rel="stylesheet" href="/css/cs.css" media="screen">', 'link')
     # replace_text_of_folder(csv_file_path, folder_need_to_change_path, 'script')
     # replace_text_of_folder(csv_file_path, folder_need_to_change_path, 'img')
-    replace_text_of_folder(csv_file_path, folder_need_to_change_path, 'link')
+    replace_text_of_folder(csv_file_path_test, folder_need_to_change_path, 'link')
